Rename container to canvas provider
