+++ blocks/motion-background/src/motion-background-canvas-provider/edit.js
